+++ db/migrate/20150618062654_increase_media_local_id_field.rb
@@ -1,0 +1,9 @@
+class IncreaseMediaLocalIdField < ActiveRecord::Migration
+  def up
+    change_column :media_items, :local_id, :string, limit: 512
+  end
+
+  def down
+    change_column :media_items, :local_id, :string, limit: 64
+  end
+end