+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 #Changelog
 
 ## latest
+- add title to media items of type voice/audio
+- change type of media items duration from int to float
+- rename voice to audio
+- rename photo to image
+- rename clyid to publicId
+
+
+## v0.5.3 on t2s2-dev
+- increase length of media item local_id from 64 to 512
+
+
+## v0.5.2 on t2s2-dev
+- unify slot media items into one array with type info for every item
+
+
+## v0.5.1 on t2s2-silvio-dev
+- fix whitelisted location attributes from snake_case to camelCase
+- re-add media items grouped by type to not make Alex depended on Antonios changes
 
 
 ## v0.5.0 on t2s2-dev